Move examples to separate mods
